Corrected actor_player sprite starting orientation in _ready()
The player sprite is facing left by default, even though the player entity thinks it is facing right. We were already compensating for this in the code, but only when the player moves. So if the player started the game and fired shots before moving, the player sprite would be facing left but the shots would go right. Now the sprite starts facing right, so the gun orientation will match.