Bug squashing
Fixed bugs that caused score to behave strangely in level 2. (Level 2's non-global score and high_score variables still exist, but are now ignored.) Resetting a level now reverts score to what it was at the start of the level (mitigates incentive to reset in order to farm more points). Score is also submitted to the leaderboard and zeroed every time the main menu is loaded (again, to deter reset farming). Fixed a bug that caused Lv2 to load if the player finished Lv2 and then chose Restart from the pause menu in Lv3. Pressing END now ends Lv2 by reducing the timer to 0.01 seconds. (Temporarily, for debugging variables across level transitions.)