Added HUDs to the rocket level
Removed the "cheat" input.
Changed the heart power up to add the heart power up's value to the player's hp (it was doubling).
Changed the player's script to restart the current scene upon death instead of freezing (still need to implement scene transitions, like fade out).
Added the health HUD to the rocket level.
Added a progress HUD to the rocket level.